Use BigDecimals for currency values

+++ maat-court-data-api/src/main/java/gov/uk/courtdata/dto/FinancialAssessmentDTO.java
@@ -5,6 +5,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -25,8 +26,8 @@ public class FinancialAssessmentDTO {
     private LocalDateTime initialAssessmentDate;
     private String initOtherBenefitNote;
     private String initOtherIncomeNote;
-    private Double initTotAggregatedIncome;
-    private Double initAdjustedIncomeValue;
+    private BigDecimal initTotAggregatedIncome;
+    private BigDecimal initAdjustedIncomeValue;
     private String initNotes;
     private String initResult;
     private String initResultReason;
@@ -40,10 +41,10 @@ public class FinancialAssessmentDTO {
     private String fullResultReason;
     private String fullAssessmentNotes;
     private String fullResult;
-    private Double fullAdjustedLivingAllowance;
-    private Double fullTotalAnnualDisposableIncome;
+    private BigDecimal fullAdjustedLivingAllowance;
+    private BigDecimal fullTotalAnnualDisposableIncome;
     private String fullOtherHousingNote;
-    private Double fullTotalAggregatedExpenses;
+    private BigDecimal fullTotalAggregatedExpenses;
     private Integer fullAscrId;
     private LocalDateTime dateCompleted;
     private LocalDateTime updated;
